Move get_current_test_name to megalinter.utils
Fixes https://github.com/oxsecurity/megalinter/issues/1685
Fixes https://github.com/oxsecurity/megalinter/issues/1684
