+++ megalinter/tests/test_megalinter/helpers/utilstest.py
@@ -604,13 +604,3 @@ def assert_file_has_been_updated(file_name, bool_val, test_self):
         test_self.assertTrue(updated, f"{file_name} has been updated")
     else:
         test_self.assertFalse(updated, f"{file_name} has not been updated")
-
-
-def get_current_test_name(full_name=False):
-    current_name = os.environ.get("PYTEST_CURRENT_TEST", None)
-    if current_name is not None:
-        if full_name is True:
-            return current_name
-        else:
-            return current_name.split(":")[-1].split(" ")[0]
-    return ""
